Finished project with extra features
This is the final project of the alura course that i was doing, with extra features that i wanted to implement for practice and curiosity purposes. Every episode in my database has two extra fields: director and synopsis. And, in order to get the data, I had to access a deeper layer of the OMDB API. This is why my "searchEpisodeBySeries" method is longer. Also implemented "getAllEpisodes" and "worstEpisodesBySeries" because i was curious about what would that be on a couple series.